modifying users

+++ app/src/routes/users.py
@@ -1,12 +1,9 @@
-import json
 from fastapi import APIRouter, Depends, Request, Response, status
 from fastapi.responses import RedirectResponse, HTMLResponse
 from fastapi.templating import Jinja2Templates
 from sqlalchemy.orm import Session
-from typing import List
-
+# Local imports
 from . helpers import get_current_user
-
 from .. db import users_actions
 from .. db.database import get_db
 from .. db.models import UsersBase
@@ -19,19 +16,12 @@ router = APIRouter(
     tags=['users']
 )
 
-@router.get('/get_users')
-def get_users(db: Session = Depends(get_db)):
-    return users_actions.get_all_users(db=db)
 
-@router.get('/get_user/{id}')
-def get_user(id:int,db: Session = Depends(get_db)):
-    return users_actions.get_user(db=db, id=id)
-
-@router.post('/create_user')
+@router.post('/users')
 async def create_user(request: Request,response:Response,db: Session = Depends(get_db)):
+    print("Trying user creation")
+    user, body = get_current_user(request=request, db=db)
     try:
-        print("Trying user creation")
-        user, body = get_current_user(request=request, db=db)
         if user is None:
             return RedirectResponse(url="/home",status_code=status.HTTP_307_TEMPORARY_REDIRECT)
         if user.rol_id != 1:
@@ -48,7 +38,7 @@ async def create_user(request: Request,response:Response,db: Session = Depends(g
         )
         userdb = users_actions.create_user(db=db,user=user)
         if isinstance(userdb, Users):
-            return RedirectResponse(url="/users/users",status_code=status.HTTP_307_TEMPORARY_REDIRECT)
+            return RedirectResponse(url="/users/users",status_code=status.HTTP_302_FOUND)
         if isinstance(userdb, str):
             body["msg"] = userdb
             return templates.TemplateResponse("users.html", body)
@@ -56,18 +46,52 @@ async def create_user(request: Request,response:Response,db: Session = Depends(g
         body["msg"] = str(e)
         return templates.TemplateResponse("users.html", body)
 
-@router.patch('/update_user')
-def update_user(user:UsersBase,db: Session = Depends(get_db)):
-    return users_actions.update_user(db=db, edit=user)
+@router.get('/update_user/{id}')
+def update_user_page(request: Request, response:Response, id:int, db: Session = Depends(get_db)):
+    print("Loading edit user page")
+    user, body = get_current_user(request=request, db=db)
+    if user is None:
+        return RedirectResponse(url="/home",status_code=status.HTTP_307_TEMPORARY_REDIRECT)
+    if user.rol_id != 1:
+        body["msg"] = "Unauthorized user getting back to home page"
+        return templates.TemplateResponse("users.html", body)
+    edit_user = db.query(Users).filter(Users.id == id).first()
+    body["user"] = edit_user
 
-@router.delete('/delete_user/{id}')
-def delete_user(id:int,db: Session = Depends(get_db)):
-    return users_actions.delete_user(db=db, id=id)
+    return templates.TemplateResponse("edit-user.html", body)
 
-@router.post('/change_roles/{id}')
-def change_roles(id:int,rol_id:int,db: Session = Depends(get_db)):
-    return users_actions.change_roles(db=db,id=id,rol_id=rol_id)
+@router.post('/update_user/{id}')
+async def update_user(request: Request, response:Response, id:int, db: Session = Depends(get_db)):
+    print("Editing user page")
+    user, body = get_current_user(request=request, db=db)
+    if user is None:
+        return RedirectResponse(url="/home",status_code=status.HTTP_307_TEMPORARY_REDIRECT)
+    if user.rol_id != 1:
+        body["msg"] = "Unauthorized user getting back to home page"
+        return templates.TemplateResponse("home.html", body)
+    edit_user = db.query(Users).filter(Users.id == id).first()
+    formdata = await request.form()
+    edit_user.name = formdata['name']
+    edit_user.last_name = formdata['last_name']
+    edit_user.email = formdata['email']
+    edit_user.password = edit_user.password if formdata['password'] =="" or formdata['password'] is None else  users_actions.Hash.bcrypt(formdata['password'])
+    edit_user.rol_id = formdata['rol']
+    updated_user = users_actions.update_user(db=db, user=user)
+    
+    return RedirectResponse(url="/users/users", status_code=status.HTTP_302_FOUND)
 
+
+@router.get('/delete_user/{id}')
+def delete_user(request: Request, response:Response, id:int,db: Session = Depends(get_db)):
+    print("Deleting user")
+    user, body = get_current_user(request=request, db=db)
+    if user is None:
+        return RedirectResponse(url="/home",status_code=status.HTTP_307_TEMPORARY_REDIRECT)
+    if user.rol_id != 1:
+        body["msg"] = "Unauthorized user getting back to home page"
+        return templates.TemplateResponse("home.html", body)
+    dleted_user = users_actions.delete_user(db=db, id=id)
+    return RedirectResponse(url="/users/users", status_code=status.HTTP_302_FOUND)
 
 
 @router.get('/users',response_class=HTMLResponse)
@@ -80,7 +104,7 @@ def get_users_page(request: Request,response:Response, db: Session = Depends(get
         print("Unauthorized user getting back to home page")
         return RedirectResponse(url="/home",status_code=status.HTTP_307_TEMPORARY_REDIRECT)
     
-    users = db.query(Users.name, Users.last_name, Users.email, Roles.name.label('rol'))\
+    users = db.query(Users.id,Users.name, Users.last_name, Users.email, Roles.name.label('rol'))\
                 .join(Roles, Roles.id == Users.rol_id)
     
     body["users"] = users
